requirements.txt
asgiref==3.7.2
Django==4.2.6
django-filter==23.3
django-polymorphic==3.1.0
djangorestframework==3.14.0
Markdown==3.5
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3